+++ src/main/java/com/ssmoker/smoker/global/configuration/SecurityConfig.java
@@ -1,6 +1,6 @@
-package com.ssmoker.smoker.security;
+package com.ssmoker.smoker.global.configuration;
 
-import com.ssmoker.smoker.security.filter.JwtRequestFilter;
+import com.ssmoker.smoker.global.security.filter.JwtRequestFilter;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Bean;
@@ -35,7 +35,7 @@ public class SecurityConfig {
             "api/smoking-area/{smokingAreaId}/rating_info", //총 별점 & 별점 개수 조회
             "api/smoking-area", //흡연구역 마커 보이기
             "api/smoking-area/{smokingAreaIid}/simple", //흡연 구역 지도 마커 클릭했을 때 흡연 구역 정보 조회 (모달)
-            "api/smokin-area/list", //흡연구역 목록보기
+            "api/smoking-area/list", //흡연구역 목록보기
             "api/smoking-area/search", //흡연 구역 검색
             "/test/{status}", //테스트 컨트롤러
             //등등 추가 및 수정해주세요
