+++ src/main/java/com/ssmoker/smoker/global/security/handler/annotation/AuthUser.java
@@ -1,5 +1,6 @@
-package com.ssmoker.smoker.security.handler.annotation;
+package com.ssmoker.smoker.global.security.handler.annotation;
 
+import io.swagger.v3.oas.annotations.Parameter;
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
@@ -7,5 +8,6 @@ import java.lang.annotation.Target;
 
 @Retention(RetentionPolicy.RUNTIME)
 @Target(ElementType.PARAMETER)
+@Parameter(hidden = true)
 public @interface AuthUser {
 }
